feat(jest): use babel-jest by default for all code
