+++ .changeset/healthy-apricots-stare.md
@@ -1,0 +1,6 @@
+---
+'arui-scripts': major
+---
+
+Теперь по умолчанию jest будет использовать babel для преобразования ts-кода. Если по какой-то причине это
+вас не устраивает - вы можете использовать настройку `jestUseTsJest`.
