+++ packages/arui-scripts/docs/settings.md
@@ -244,3 +244,6 @@ const settings = {
 Вместо этого будет запускаться отдельный процесс `tsc`, который будет выдавать сообщения об ошибках в типах только в консоль.
 
 По умолчанию `true`.
+
+#### jestUseTsJest
+Позволяет использовать `ts-jest` вместо `babel-jest` для обработки `ts` файлов при запуске тестов. По умолчанию `false`.
